core: Renamed qtlab_flow to flow.
